.sql data cleaning script upload
This script is centered around the data cleaning process. We conducted data cleaning on the world_life_expectancy table using the MYSQL platform. We encountered some challenges that required us to implement workarounds to achieve the desired outcome. Throughout this project, the course instructor provided a follow-up video in which I, as a student, worked alongside the instructor.